requirements.txt
streamlit>=1.32.0
plaid-python>=14.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
flask>=3.0.0
